requirements.txt
groq
langchain
langchain-groq
langchain-community
ddgs
pydantic
python-dotenv
colorama
httpx
